Add inital undocumented mackup add command
