+++ internal/cmd/addcmd.go
@@ -102,7 +102,10 @@ func (c *Config) defaultPreAddFunc(targetRelPath chezmoi.RelPath, newSourceState
 }
 
 func (c *Config) runAddCmd(cmd *cobra.Command, args []string, sourceState *chezmoi.SourceState) error {
-	destAbsPathInfos, err := c.destAbsPathInfos(sourceState, args, c.Add.recursive, c.Add.follow)
+	destAbsPathInfos, err := c.destAbsPathInfos(sourceState, args, destAbsPathInfosOptions{
+		follow:    c.Add.follow,
+		recursive: c.Add.recursive,
+	})
 	if err != nil {
 		return err
 	}
