+++ internal/cmd/noupgradecmd.go
@@ -3,12 +3,11 @@
 package cmd
 
 import (
-	"fmt"
-	"runtime"
-
 	"github.com/spf13/cobra"
 )
 
-func (c *Config) runUpgradeCmd(cmd *cobra.Command, args []string) error {
-	return fmt.Errorf("%s: unsupported GOOS", runtime.GOOS)
+type upgradeCmdConfig struct{}
+
+func (c *Config) newUpgradeCmd() *cobra.Command {
+	return nil
 }
